+++ Apps.Hubspot/Actions/MarketingEmailsActions.cs
@@ -21,6 +21,7 @@ using Blackbird.Applications.Sdk.Common.Authentication;
 using Blackbird.Applications.Sdk.Utils.Html.Extensions;
 using HtmlAgilityPack;
 using Blackbird.Applications.Sdk.Common.Exceptions;
+using Apps.Hubspot.Utils.Extensions;
 
 namespace Apps.Hubspot.Actions;
 
@@ -97,7 +98,7 @@ public class MarketingEmailsActions(InvocationContext invocationContext, IFileMa
         var htmlDoc = new HtmlDocument();
         htmlDoc.Load(htmlFile);
 
-        var extractedValues = ExtractHtmlValues(htmlDoc);
+        var extractedValues = Apps.Hubspot.Utils.Extensions.HtmlExtensions.ExtractHtmlValues(htmlDoc);
 
         var createRequest = new CreateMarketingEmailOptionalRequest
         {
@@ -116,44 +117,6 @@ public class MarketingEmailsActions(InvocationContext invocationContext, IFileMa
 
         return await Client.ExecuteWithErrorHandling<MarketingEmailDto>(request); 
     }
-
-
-    private HtmlValues ExtractHtmlValues(HtmlDocument htmlDoc)
-    {
-        var titleNode = htmlDoc.DocumentNode.SelectSingleNode("//title");
-        var title = titleNode?.InnerText.Trim()
-            ?? throw new PluginApplicationException("The HTML file does not contain a valid title.");
-        
-        var bodyNode = htmlDoc.DocumentNode.SelectSingleNode("//body");
-        if (bodyNode == null)
-        {
-            throw new PluginApplicationException("The HTML file does not contain a valid body section.");
-        }
-
-        var nameNode = htmlDoc.DocumentNode.SelectSingleNode("//*[@id='name']");
-        var subjectNode = htmlDoc.DocumentNode.SelectSingleNode("//*[@id='subject']");
-        var sendOnPublishNode = htmlDoc.DocumentNode.SelectSingleNode("//*[@id='sendOnPublish']");
-        var archivedNode = htmlDoc.DocumentNode.SelectSingleNode("//*[@id='archived']");
-        var activeDomainNode = htmlDoc.DocumentNode.SelectSingleNode("//*[@id='activeDomain']");
-        var languageNode = htmlDoc.DocumentNode.SelectSingleNode("//*[@id='language']");
-        var publishDateNode = htmlDoc.DocumentNode.SelectSingleNode("//*[@id='publishDate']");
-        var businessUnitIdNode = htmlDoc.DocumentNode.SelectSingleNode("//*[@id='businessUnitId']");
-
-        return new HtmlValues
-        {
-            Title = title,
-            Body = bodyNode.InnerHtml,
-            Name = nameNode?.InnerText.Trim(),
-            Subject = subjectNode?.InnerText.Trim(),
-            SendOnPublish = sendOnPublishNode != null && bool.TryParse(sendOnPublishNode.InnerText.Trim(), out var sendOnPublish) ? sendOnPublish : (bool?)null,
-            Archived = archivedNode != null && bool.TryParse(archivedNode.InnerText.Trim(), out var archived) ? archived : (bool?)null,
-            ActiveDomain = activeDomainNode?.InnerText.Trim(),
-            Language = languageNode?.InnerText.Trim(),
-            PublishDate = publishDateNode != null && DateTime.TryParse(publishDateNode.InnerText.Trim(), out var publishDate) ? publishDate : (DateTime?)null,
-            BusinessUnitId = businessUnitIdNode?.InnerText.Trim()
-        };
-    }
-
 
     private Task<EmailContentDto> GetEmail(string emailId)
     {
